clippy: clear the warning backlog
835 clippy warnings had accumulated (we'd only ever been checking for denied
*errors*, so default-level warnings went unnoticed). Cleared:

- 749 were `excessive_precision` on canonical constant tables (libopus
  twiddle/window tables, energy means, tapset gains, preemph). Kept verbatim
  - they round to the correct f32 and trimming loses reference provenance -
  with `allow` scoped to the data: module-level on the all-tables file,
  item-level on the three inline consts.
- ~75 real style nits fixed: needless_borrow, parenthesized precedence
  (semantics-preserving), inlined late-inits, redundant `as i32` casts (left
  over from the c_int->i32 pass), manual_memcpy -> copy_from_slice, and an
  if/else with identical wideband/else arms collapsed (matches the C).
- `neg_cmp_op_on_partial_ord`: kept `!(s1 > 0.2*s2)` with a scoped allow -
  the negation is deliberate so an unordered/NaN compare counts as an
  explosion; `<=` would change NaN behavior.
- `needless_range_loop` allowed only in true kernels (cwrs), the opus_compare
  dev bin, and tests. The last `vec!` (a cwrs test) became a fixed array,
  which also fixes the no_std test build.

Float+std, fixed+std and fixed+no_std are warning-clean; bit-exact vectors;
37 tests pass. (The libm build's remaining FloatMath warnings are handled in
the next commit.)

+++ src/packet.rs
@@ -339,7 +339,7 @@ mod tests {
         // Code 0 = one frame, CELT 2.5ms NB (TOC = 0x80)
         let pkt = [0x80u8, 0xAA, 0xBB, 0xCC];
         let mut toc: u8 = 0;
-        let mut offsets = [0 as i32; 48];
+        let mut offsets = [0_i32; 48];
         let mut sizes = [0i16; 48];
         let mut offset: i32 = 0;
 
@@ -355,7 +355,7 @@ mod tests {
         // Code 1 = two CBR frames, CELT 2.5ms NB; 6 bytes payload -> 3 each
         let pkt = [0x81u8, 1, 2, 3, 4, 5, 6];
         let mut toc: u8 = 0;
-        let mut offsets = [0 as i32; 48];
+        let mut offsets = [0_i32; 48];
         let mut sizes = [0i16; 48];
         let mut offset: i32 = 0;
 
@@ -381,7 +381,7 @@ mod tests {
         // Code 2 = two VBR frames; size[0] = 2 (single-byte), remaining = last
         let pkt = [0x82u8, 2, 0xAA, 0xBB, 0xCC, 0xDD, 0xEE];
         let mut toc: u8 = 0;
-        let mut offsets = [0 as i32; 48];
+        let mut offsets = [0_i32; 48];
         let mut sizes = [0i16; 48];
         let mut offset: i32 = 0;
